# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Kasir API",
        "description": "Dokumentasi API untuk aplikasi kasir.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "API base URL"
        }
    ],
    "paths": {
...
        "/produk": {
            "get": {
                "tags": [
                    "Produk"
                ],
                "summary": "Daftar produk",
                "operationId": "b5bb4927f9ba0bb6ca925c3abaf93ce4",
                "parameters": [
                    {
                        "name": "q",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "kategori_id",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Daftar produk berhasil diambil"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Produk"
                ],
                "summary": "Tambah produk",
                "operationId": "b19ea02c062be79eb7a0cc6e9bc24d70",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nama",
                                    "kategori_id",
                                    "harga",
                                    "stok",
                                    "kode_produk"
                                ],
                                "properties": {
                                    "nama": {
                                        "type": "string",
                                        "example": "Teh Botol"
                                    },
                                    "kategori_id": {
                                        "type": "string",
                                        "example": "01kty..."
                                    },
                                    "harga": {
                                        "type": "integer",
                                        "example": 5000
                                    },
                                    "stok": {
                                        "type": "integer",
                                        "example": 20
                                    },
                                    "kode_produk": {
                                        "type": "string",
                                        "example": "PRD-001"
                                    },
                                    "is_active": {
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Produk berhasil dibuat"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "422": {
                        "description": "Validasi gagal"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/produk/{produk}": {
            "get": {
                "tags": [
                    "Produk"
                ],
                "summary": "Detail produk",
                "operationId": "8e04f455a07dc5d65ba8ba8f094a557d",
                "parameters": [
                    {
                        "name": "produk",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detail produk berhasil diambil"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "Produk tidak ditemukan"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Produk"
                ],
                "summary": "Ubah produk",
                "operationId": "2a621274dc23f399dcc9ec26993118a7",
                "parameters": [
                    {
                        "name": "produk",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nama": {
                                        "type": "string",
                                        "example": "Teh Botol Sosro"
                                    },
                                    "kategori_id": {
                                        "type": "string",
                                        "example": "01kty..."
                                    },
                                    "harga": {
                                        "type": "integer",
                                        "example": 6000
                                    },
                                    "stok": {
                                        "type": "integer",
                                        "example": 25
                                    },
                                    "kode_produk": {
                                        "type": "string",
                                        "example": "PRD-001"
                                    },
                                    "is_active": {
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Produk berhasil diubah"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Produk tidak ditemukan"
                    },
                    "422": {
                        "description": "Validasi gagal"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Produk"
                ],
                "summary": "Hapus produk",
                "operationId": "19fdc1eec8aefe5f939df26f2f9d1ced",
                "parameters": [
                    {
                        "name": "produk",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Produk berhasil dihapus"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Produk tidak ditemukan"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Masukkan bearer token hasil login.",
                "bearerFormat": "Token",
                "scheme": "bearer"
            }
        }
    },
...
}